// src/lib/cv/types.ts
export interface CVSettings {
  maxPagesAllowed: number;
  allowThreePages: boolean;
  isHarvardTemplate: boolean;
  showPageLimitWarning: boolean;
  trimmedItems: {
    experiences: number;
    education: number;
    certifications: number;
    references: number;
  };
}